requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-genai>=0.2.0
